chore: Add eslint rule for Promises (fix #2065)
Adds returns to all Promises and ensures all test code that
uses a Promise returns it so all code is executed.

This is a lot of noise but and makes for slightly more verbose
tests with "needless" `return true;` statements, but it will
prevent us from writing unexecuted test code in the future. 👍🏻